Move ManagerProxyBinder to DLEDotNet.Data; seems to fit better there
